pipeline: nuke cache.qcow2 once it grows past 8G
Otherwise, we'll just hit ENOSPC eventually. rpm-ostree doesn't/can't
know what it should delete since (not surprisingly) the cache is meant
to cache packages that may be used in the next compose. We could try to
be smart and e.g. only keep the latest NEVRA of all packages, etc... but
really it's just easier to let rpm-ostree re-download and import it all
once in a while. It's good to exercise the import code too since it can
subtly change between rpm-ostree versions.