Add support for developer pipelines
So, a lot happening in this patch, but essentially the idea is:

1. add some more parameters to the template to support "devel" resources
2. add annotations to the pipeline buildconfig to allow us to pass data
   to the pipeline code; update the pipeline to pick this up
3. allow uploading to S3 in devel mode using the provided devel prefix
4. add a new `devel-up` utility to make it easier and safer to create
   and update devel pipelines; no way to delete them yet, will do that
   in a follow up

The reason for the `devel-up` utility is that I wanted more "semantic"
control over the way devel resources are created than through directly
using parameters, especially around ensuring that they absolutely do not
modify prod resources. But also, it's much more convenient to use than
overriding each param one by one.