Add support for pushing extensions container and legacy oscontainer
Those are two containers we need to push in RHCOS that's not in FCOS.
Expand the list of supported keys in `registry_repos` for those and
accordingly rework the logic in the release job to be more functional.

Make it automatically skip containers that weren't actually built (e.g.
`extensions` are not built on FCOS and RHCOS <= 4.11).

Support stream-specific additional `registry_repos` so that RHCOS <=
4.11 can specify the `registry.ci.openshift.org` location, which is no
longer needed in 4.12.

Add a new `add_build_tag` which is required to ensure that the pushed
images are not garbage-collected.

Use the new `cosa copy-container` which knows to "unmanifest list" when
copying from Quay.io to `registry.ci.openshift.org`.