Replace cosa controller image heuristic with temporary pipecfg knob
We have this hack in the `build-arch` job to work around the fact that
the remote code doesn't exist in branched cosa. It uses a heuristic
based on the tag name to know when to use the latest cosa as the
controller knob rather than respecting the configured cosa image. This
is really subtle and prone to error.

Let's instead rework this so that it's driven from the pipecfg as a
stream-specific key. Since the hack is only necessary for 4.11 and
older, we mark it as temporary. As we shed older streams from the
config, the key will be used less often, until the point 4.11 goes EOL
and we can drop support for it entirely. (This is similar to our
approach to `skip_artifacts`.)