jobs/build: drop `official` gate on release job trigger
We're moving to a model where if the right credentials exist, we assume
that the pipeline is permitted to affect the outside world (e.g. push
containers, import OSTree commits, etc...).

In a future commit, we'll add safeguards to ensure that the canonical
external endpoints cannot be pushed to except from the official Jenkins.
But we don't need to dynamically support both ways in the pipeline.