Add `versionary_hack` knob
I'd like to get rid of the `official` variable, and this is the final
user of it.

For now, hack around this by adding another config knob. In the future,
I think there's different ways we could tackle this. The simplest would
be to move the versionary script into the pipecfg repo, and then the
pipeline automatically uses that script if it's found (though it would
require an extra `checkout scm`). Another approach is lowering this into
cosa/rpm-ostree and f-c-c somehow, but it needs some design work. It
depends also if we want to keep the property of developer builds keeping
the `dev` marker.