Add basic support for hotfix builds
ART needs the ability to perform hotfix builds, i.e. builds outside the
regular stream for a specific purpose. These builds must not pollute
production stream references.

Support this via the new `hotfix` section of the pipecfg. Require a
`name` subkey which is then used to modify the S3 subpath we upload to
and the oscontainer tags we push to. Add new hotfix parameters to the
`build`, `build-arch`, and `release` jobs which will be used to pass the
repo containing the hotfix pipecfg.

Don't support hotfix cloud uploads since it's not yet needed and
requires plumbing of the hotfix prefix.