jobs/release: flip stage scope
Let's only add a scope if we actually enter the
`tryWithMessagingCredentials` block. That way we don't have a phantom
scope in e.g. the RHCOS pipeline where we don't sign.

And since we do that, we might as well simplify and merge with the very
next block which was also conditional on messaging credentials.